rearrange files and fixed harmony prefix issue
